more fun debug casts

+++ bp.cpp
@@ -195,9 +195,9 @@ public:
         uint8_t fsm_sub_index = hist ^ parts.share_bits;
 
         std::cout << "fsms.size(): " << fsms.size()
-            << "; fsm_index: " << fsm_index
+            << "; fsm_index: " << static_cast<int>(fsm_index)
             << "; fsms[fsm_index].size(): " << fsms[fsm_index].size()
-            << "; fsm_sub_index: " << fsm_sub_index
+            << "; fsm_sub_index: " << static_cast<int>(fsm_sub_index)
             << "; fsms[fsm_index][fsm_sub_index]: " << fsms[fsm_index][fsm_sub_index]
             << std::endl;
